refactor: Update application namespace for android
- Increase version to 2.0.1
