+++ android/app/src/main/java/dev/nedaa/android/MainApplication.kt
@@ -1,4 +1,4 @@
-package dev.nedaa.app
+package dev.nedaa.android
 
 import android.app.Application
 import android.content.res.Configuration
